Added event listen function.
